refactor: close ADR-061 Phase 2's compare -> policy blocker
Re-measuring the blocker ADR-061 Phase 2 recorded found one real forbidden
edge rather than the five it named: checker_types.py's function-local
`from .severity import effective_verdict_for_change`. That single edge also
closed the `compare -> policy -> compare` cycle against analysis_assurance's
own (allowed) `policy -> compare` import of DiffResult. The other four cited
positions never fired -- checker.py is not classified at all, and
checker_types.py's contract_gating imports only violate the contract if
contract_gating.py is itself classified `policy`, which this change
deliberately declines to do.

Remove that edge by extracting the shared logic to a leaf both layers may
depend on -- the same pattern ADR-061's Phase 3 blocker note names.
`effective_verdict_for_change`, its disclosure sibling
`reclassify_rule_for_change`, and the `KindSets`/`resolve_kind_sets` pair
they share move from severity.py into reclassify.py, which already owned the
selector-scoped rules that resolver's precedence chain is built around (and
whose two mirrored implementations of that chain had already needed three
review rounds to stop disagreeing).

severity.py re-exports all three through the redundant-alias form, so
`abicheck.severity.effective_verdict_for_change` and its siblings keep
working unchanged for every caller in and out of this repo, and DiffResult's
public breaking/source_breaks/compatible/risk properties are untouched.

With the edge gone, severity.py and analysis_assurance.py are classified
`policy` in architecture/modules.yaml and check_architecture.py reports zero
findings. reclassify.py and contract_gating.py stay deliberately unclassified
-- they are the leaves both `compare`'s result type and `policy`'s severity
layer depend on, and which layer owns them is decided by checker_policy.py's
own model-vs-policy split, not by this slice.
